Addressed PR Review comments
-- Cleaned up testing code
-- Clarified and cleaned up param documentation
-- Moved the config validation into initServer
