+++ config/config.go
@@ -2117,6 +2117,10 @@ func InitServer(ctx context.Context, currentServers server_structs.ServerType) e
 		return err
 	}
 
+	if err := ValidateLogExportsConfig(); err != nil {
+		return err
+	}
+
 	return nil
 }
