+++ docs/parameters.yaml
@@ -438,7 +438,8 @@ name: Logging.LogExports.Enabled
 description: |+
   When true, this server exports its logs via the Pelican logging namespace (`/pelican/logging/{Server.ID}`).
   Clients with the appropriate token can retrieve log entries as virtual time-bucketed objects.
-  Requires `Logging.LogLocation` to be set to a real file path (not empty or `/dev/null`).
+  Requires `${Logging.LogLocation}` to be set to a real file path (not empty or `/dev/null`).
+  `${IssuerKeysDirectory}` must also be configured so the server can auto-generate access tokens for the logging namespace.
   Defaults to false — log export is opt-in.
 type: bool
 default: false
@@ -448,7 +449,7 @@ name: Logging.LogExports.AllowFederationAdmin
 description: |+
   When true, the federation issuer URL is added as an authorized issuer for this server's logging namespace,
   allowing federation administrators to read logs from this server using a federation-issued token.
-  Has no effect when `Logging.LogExports.Enabled` is false.
+  Has no effect when `${Logging.LogExports.Enabled}` is false.
   Defaults to false.
 type: bool
 default: false
@@ -2960,7 +2961,7 @@ description: |+
   submitted by Origins or Caches whose main server registration is already approved.
   If the server registration is still pending at the time the logging namespace is submitted,
   approval is deferred until the server is approved.
-  This is a federation-level gate distinct from `Logging.LogExports.Enabled` on each individual server.
+  This is a federation-level gate distinct from `${Logging.LogExports.Enabled}` on each individual server.
   Defaults to false.
 type: bool
 default: false
